use ajax to submit the announcement info and fix the resource'ssavepath

+++ ccweb/src/main/java/com/knet51/ccweb/util/fileUpLoad/SavePathUtil.java
@@ -3,7 +3,7 @@ package com.knet51.ccweb.util.fileUpLoad;
 import java.io.File;
 
 public class SavePathUtil {
-	public static String getPath(String fileName,Integer id,String type){
+	public static String getPath(String fileName,Long id,String type){
 		String path = "d:\\"+fileName+"\\"+id+"\\"+type;
 		File f = new File(path);
 		if(!f.exists()) {
